pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "patchtest2"
dynamic = ["version"]
description = 'Modern refactor of the Yocto Project patchtest tool for validating patches'
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = [
    "LICENSE.txt",
]
keywords = ["patch", "testing", "yocto", "openembedded"]
authors = [
  { name = "Trevor Gamblin", email = "tgamblin@baylibre.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "GitPython",
  "pyparsing",
  "unidiff",
]

[project.optional-dependencies]
dev = [
  "black>=23.0.0",
  "isort>=5.12.0",
  "mypy>=1.0.0",
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "types-setuptools",
  "types-unidiff",
]
test = [
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
]

[project.scripts]
patchtest = "patchtest2.patchtest:run"

[project.urls]
Documentation = "https://github.com/threexc/patchtest2#readme"
Issues = "https://github.com/threexc/patchtest2/issues"
Source = "https://github.com/threexc/patchtest2"

[tool.hatch.version]
path = "src/patchtest2/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/patchtest2 tests}"

[tool.coverage.run]
source_pkgs = ["patchtest2", "tests"]
branch = true
parallel = true
omit = [
  "src/patchtest2/__about__.py",
]

[tool.coverage.paths]
patchtest2 = ["src/patchtest2", "*/patchtest2/src/patchtest2"]
tests = ["tests", "*/patchtest2/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.black]
line-length = 88
target-version = ["py310"]
include = "\\.pyi?$"
extend-exclude = """
/(
    \\.eggs
  | \\.git
  | \\.hg
  | \\.mypy_cache
  | \\.tox
  | \\.venv
  | _build
  | buck-out
  | build
  | dist
)/
"""

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["patchtest2"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.hatch.build.targets.wheel]
packages = ["src/patchtest2"]
